// frontend/src/app/shared/generated-api/model/invitationCreateRequest.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InvitationCreateRequest { 
    email: string;
    proposedUsername?: string;
    displayName?: string;
    role: InvitationCreateRequestRoleEnum;
}
export enum InvitationCreateRequestRoleEnum {
    User = 'USER',
    Admin = 'ADMIN'
};