// frontend/src/app/shared/generated-api/model/invitationDelivery.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InvitationSummary } from './invitationSummary';


export interface InvitationDelivery { 
    invitation: InvitationSummary;
    delivery: InvitationDeliveryDeliveryEnum;
    token?: string;
    activationUrl?: string;
}
export enum InvitationDeliveryDeliveryEnum {
    EmailSent = 'EMAIL_SENT',
    Manual = 'MANUAL'
};