// frontend/src/app/shared/generated-api/model/invitationSummary.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InvitationSummary { 
    id: string;
    email: string;
    proposedUsername?: string;
    displayName?: string;
    role: InvitationSummaryRoleEnum;
    status: InvitationSummaryStatusEnum;
    createdAt: string;
    expiresAt: string;
    acceptedAt?: string;
    revokedAt?: string;
}
export enum InvitationSummaryRoleEnum {
    User = 'USER',
    Admin = 'ADMIN'
};
export enum InvitationSummaryStatusEnum {
    Pending = 'PENDING',
    Accepted = 'ACCEPTED',
    Revoked = 'REVOKED',
    Expired = 'EXPIRED'
};